# api/openapi/template.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CountdownServer",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [],
  "paths": {
    "/templates": {
      "post": {
        "summary": "创建邮件模板",
        "deprecated": false,
        "description": "",
        "operationId": "createTemplate",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "topic": {
                    "type": "string",
                    "title": "邮件主题",
                    "maxLength": 255
                  },
                  "content": {
                    "type": "string",
                    "title": "邮件内容",
                    "maxLength": 65535
                  }
                },
                "required": [
                  "topic",
                  "content"
                ]
              },
              "example": {
                "type": 0,
                "topic": "string",
                "content": "string",
                "slots": [
                  "string"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数不正确",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": []
      },
      "get": {
        "summary": "通用查询接口",
        "deprecated": false,
        "description": "",
        "operationId": "getTemplates",
        "tags": [],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "模板类型",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "contentLen",
            "in": "query",
            "description": "界面中显示的content长度",
            "required": false,
            "example": "255",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "$ref": "#/components/schemas/template"
                    }
                  },
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数不正确",
            "content": {
              "application/json": {
                "schema": {
                  "title": "",
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/templates/{tid}": {
      "get": {
        "summary": "根据id查询邮件模板",
        "deprecated": false,
        "description": "",
        "operationId": "getTemplateById",
        "tags": [],
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "data": {
                      "$ref": "#/components/schemas/template"
                    }
                  },
                  "required": [
                    "code",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数不正确",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "根据id修改模板内容",
        "deprecated": false,
        "description": "",
        "operationId": "updateTemplate",
        "tags": [],
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "topic": {
                    "type": "string",
                    "title": "邮件主题",
                    "maxLength": 255,
                    "nullable": true
                  },
                  "content": {
                    "type": "string",
                    "title": "邮件内容",
                    "maxLength": 65535,
                    "nullable": true
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": []
      },
      "delete": {
        "summary": "根据id删除模板",
        "deprecated": false,
        "description": "",
        "operationId": "deleteTemplateById",
        "tags": [],
        "parameters": [
          {
            "name": "tid",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "记录不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "template": {
        "type": "object",
        "properties": {
          "type": {
            "type": "integer",
            "description": "1不完整需填充数据2完整",
            "title": "模板类型"
          },
          "topic": {
            "type": "string",
            "title": "邮件主题"
          },
          "content": {
            "type": "string",
            "title": "邮件内容"
          },
          "slots": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "模板中需要填充的数据",
            "title": "插槽"
          },
          "updateTime": {
            "type": "string",
            "title": "更新时间",
            "format": "date-time"
          }
        },
        "required": [
          "type",
          "topic",
          "content",
          "slots",
          "updateTime"
        ]
      }
    },
    "securitySchemes": {}
  },
  "servers": []
}